Some refactoring and comments
